pyproject.toml
[tool.poetry]
name = "Medida Derivas"
version = "0.1.0"
description = ""
authors = ["Jorge Calatayud <jorge.cmaeso@upm.es>"]

[tool.poetry.dependencies]
python = ">=3.7,<3.11"
PyVISA = "^1.12.0"
numpy = "1.21.4"
scipy = "1.7.3"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
